Move delegations iterator to internal package
